Rename date tests
